+++ plugins/on-demand-job-runner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.4
+* Added kyverno policies
+
+## 0.1.3
+* Fixing vulnerability
+
 ## 0.1.2
 * Bumped go to 1.24.6 for fixing vulnerability