+++ plugins/trivy/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.33.4
+* Bump trivy to 0.67.0
+
 ## 0.33.3
 * Bump trivy to 0.66.0